merge master

+++ src/librbd/WatchNotifyTypes.h
@@ -410,10 +410,10 @@ struct MigratePayload : public AsyncRequestPayloadBase {
 };
 
 struct SparsifyPayload : public AsyncRequestPayloadBase {
-  size_t sparse_size = 0;
+  uint64_t sparse_size = 0;
 
   SparsifyPayload() {}
-  SparsifyPayload(const AsyncRequestId &id, size_t sparse_size)
+  SparsifyPayload(const AsyncRequestId &id, uint64_t sparse_size)
     : AsyncRequestPayloadBase(id), sparse_size(sparse_size) {
   }
